Moved analyzer logic to Analysis
Also Removed the Analyzer module and pushed up all the submodules to Analysis
